Enable install options for pip-installed global packages
Reason:
The '--upgrade' default option does not work well with the 'howdoi' package install as this package needs lxml, which crashes when installed with pip on cygwin.
So we cannot use the '--upgrade' option for this package, since in order to have lxml installed on cygwin, we need to install the slighly outdated cygwin package 'python3-lxml'
that comes from the cygwin package manager.
Thus a 'pip install --upgrade howdoi' would force pip to install the latest lxml version over the cygwin one, which would crash as explained above.